+++ demo/modal/footer.tsx
@@ -13,7 +13,21 @@ export default function () {
       <Button type="primary" onClick={() => setOpen(true)}>
         Modal
       </Button>
-      <Modal title="这是一个标题" open={open} onClose={onClose} onOk={onOk}>
+      <Modal
+        title="这是一个标题"
+        open={open}
+        onClose={onClose}
+        footer={
+          <div className="flex gap-2 justify-end">
+            <Button type="warning" onClick={onClose}>
+              自定义1
+            </Button>
+            <Button type="danger" onClick={onOk}>
+              自定义2
+            </Button>
+          </div>
+        }
+      >
         <div>
           Lorem ipsum dolor sit amet consectetur adipisicing elit. Iusto,
           assumenda corporis quas accusamus nesciunt quasi quisquam quibusdam